Rename train_model.train_model -> train_model.convnet
