+++ lab/processes/pytorch_example/train_model/convnet.py
